Fix bugs in tensor product benchmark

+++ cpu_benchmarks/benchmarks/naive_substring/substring.h
@@ -5,20 +5,20 @@ extern "C" {
 #endif
 struct Char {
     char c;
-} __attribute((packed, staged));
+} __attribute((packed,staged));
 
 /// Basic non-specialized function
-int find_substring(
+long long find_substring(
         const struct Char *data, 
-        int data_length, 
+        long long data_length, 
         const struct Char *pattern, 
-        int pattern_length);
+        long long pattern_length);
 
 /// Generates specialized function on pattern
 void *mix_find_substring(
         void *context,
         const struct Char *pattern,
-        int pattern_length);
+        long long pattern_length);
 
 #ifdef __cplusplus
 }
